Make sure the thread is still alive while waiting for pid in Process#kill
* Handle no pid in cleanup.

+++ spec/ruby/core/process/fixtures/common.rb
@@ -45,8 +45,8 @@ module ProcessSpecs
         args = [@pid_file, scenario, ruby_exe]
         @result = ruby_exe @script, args: args
       end
-      Thread.pass until File.exist? @pid_file
-      while @pid.nil? || @pid == 0
+      Thread.pass while @thread.status and !File.exist?(@pid_file)
+      while @thread.status && (@pid.nil? || @pid == 0)
         @pid = IO.read(@pid_file).chomp.to_i
       end
     end
@@ -54,7 +54,7 @@ module ProcessSpecs
     def wait_on_result
       # Ensure the process exits
       begin
-        Process.kill :TERM, pid
+        Process.kill :TERM, pid if pid
       rescue Errno::ESRCH
         # Ignore the process not existing
       end
